better dockerfile
