+++ flask_flights_mc/app.py
@@ -1,0 +1,13 @@
+from flask import Flask, request
+app = Flask(__name__)
+
+
+@app.route('/airports', methods = ['POST', 'GET'])
+def getAirports():
+    if request.method == 'GET':
+        return "<h1>Return a JSONified object of all airports from db coming from mysql docker container</h1>"
+    if request.method == 'POST':
+        return "<h1>Take an object of a airport. Add it to shared db instance.</h1>"
+
+if __name__ == "__main__":
+    app.run(host ='0.0.0.0', debug = True)
